Fix prisoner search tests and split out prisoner profile
